move mock_agent into its own file
